requirements.txt
Flask==0.12.2
flask-mongoengine==0.9.3
mongoengine==0.13.0
pytest-flask==0.10.0
requests==2.14.2
celery==4.0.2